改善数据结构

+++ app.py
@@ -1,26 +1,18 @@
 from flask import Flask, request, make_response
 from pathlib import Path
-from tools.general import is_allowed_ext,get_thumbnail_pic,get_tag,HOST,PathDict
+from tools.general import is_allowed_ext,get_thumbnail_pic,get_tag,\
+    HOST,PathDict,TagGroup,Tag,names,webpath_from_relpath
 
 import os,pickle,sqlite3
 from flask_cors import CORS
 
 from retrieval.img_retrieval import retrievalapp
-from detection.img_detect import detectapp
+from detects.img_detect import detectapp
 app = Flask(__name__)
 CORS(app,supports_credentials=True)
 app.register_blueprint(retrievalapp,url_prefix='/retrieval')
 app.register_blueprint(detectapp,url_prefix='/detect')
 
-names= ['person', 'bicycle', 'car', 'motorcycle', 'airplane', 'bus', 'train', 'truck', 'boat', 'traffic light',
-        'fire hydrant', 'stop sign', 'parking meter', 'bench', 'bird', 'cat', 'dog', 'horse', 'sheep', 'cow',
-        'elephant', 'bear', 'zebra', 'giraffe', 'backpack', 'umbrella', 'handbag', 'tie', 'suitcase', 'frisbee',
-        'skis', 'snowboard', 'sports ball', 'kite', 'baseball bat', 'baseball glove', 'skateboard', 'surfboard',
-        'tennis racket', 'bottle', 'wine glass', 'cup', 'fork', 'knife', 'spoon', 'bowl', 'banana', 'apple',
-        'sandwich', 'orange', 'broccoli', 'carrot', 'hot dog', 'pizza', 'donut', 'cake', 'chair', 'couch',
-        'potted plant', 'bed', 'dining table', 'toilet', 'tv', 'laptop', 'mouse', 'remote', 'keyboard', 'cell phone',
-        'microwave', 'oven', 'toaster', 'sink', 'refrigerator', 'book', 'clock', 'vase', 'scissors', 'teddy bear',
-        'hair drier', 'toothbrush']
 
 
 
@@ -28,12 +20,13 @@ names= ['person', 'bicycle', 'car', 'motorcycle', 'airplane', 'bus', 'train', 't
 #获取、查看、删除图片等基本接口在这里
 
 #获取指定文件夹下的图片
-@app.route('/get_pics/<path:base>',methods=['GET'])
-def get_pics(base):
-    if not base in PathDict:
+@app.route('/get_pics/<path:webdir>',methods=['GET'])
+def get_pics(webdir,baseindex=0):
+    if not webdir in PathDict:
         print('dir not exist')
         return {'total':0,'imgs':[]}
-    root = PathDict[base]
+    root = PathDict[webdir]
+    baseindex#用于get_all_pics的index矫正
 
     detect = sqlite3.connect("detect_results.db")#连接数据库
     cursor = detect.cursor()
@@ -46,11 +39,11 @@ def get_pics(base):
             if(not os.path.exists(thumb_rel_path)):#缩略图不存在,生成缩略图
                 get_thumbnail_pic(root+'/'+file)
 
-            tag = get_tag(cursor,base+'/'+file)
+            tag = get_tag(webdir+'/'+file)
             img = {'id':root+'/'+file,
-                   'index':num,
-                   'thumbnail':HOST + base + '/.thumbnail/' + file,
-                   'original':HOST+base+'/'+file,
+                   'index':num+baseindex,
+                   'thumbnail':HOST + webdir + '/.thumbnail/' + file,
+                   'original':HOST+webdir+'/'+file,
                    # 'webformatURL': HOST+'data/images/'+'IMG20170819123559.jpg',
                    'tags':tag}
             hits.append(img)
@@ -65,7 +58,7 @@ def get_all_pics():
     total = 0
     imgs = []
     for web_dir in PathDict:
-        res = get_pics(web_dir)
+        res = get_pics(web_dir,len(imgs))
         total += res['total']
         imgs += res['imgs']
     return {'total':total,'imgs':imgs}
@@ -87,20 +80,21 @@ def show_photo(dir,file):
 @app.route('/search/<path:tag_name>',methods=['GET'])
 def search(tag_name):
     print(tag_name)
-    if(tag_name==''): return get_pics('pics')
-    detect = sqlite3.connect("detect_results.db")
-    cursor = detect.cursor()
-    if(not tag_name in names):#没有该标签
-        return {'imgs':[]}
-    tag = names.index(tag_name)
-    cursor.execute("""select * from tagclassified where tag = ?""",(tag,))
-    result = cursor.fetchone()
-    if(result==None):#没有记录
-        return {'imgs':[]}
-    t,imgs_dump = result
-    imgs = pickle.loads(imgs_dump)
-    imgs_pack = []
+    if(tag_name==''): return get_all_pics() #返回所有图片
+    # detect = sqlite3.connect("detect_results.db")
+    # cursor = detect.cursor()
+    # if(not tag_name in names):#没有该标签
+    #     return {'imgs':[]}
+    # tag = names.index(tag_name)
+    # cursor.execute("""select * from tagclassified where tag = ?""",(tag,))
+    # result = cursor.fetchone()
+    # if(result==None):#没有记录
+    #     return {'imgs':[]}
+    # t,imgs_dump = result
+    # imgs = pickle.loads(imgs_dump)
+    imgs = TagGroup[names.index(tag_name)]
 
+    imgs_pack = []
     num = 0
     for img in imgs:
         img_splited = img.rsplit('/',1)
@@ -114,37 +108,37 @@ def search(tag_name):
               'index': num,
               'thumbnail': HOST + thumb_img,
               'original': HOST + img,
-              'tags': get_tag(cursor,img)}
+              'tags': get_tag(img)}
         imgs_pack.append(im)
         num+=1
-    detect.commit()
-    detect.close()
     return {'total':num,'imgs':imgs_pack}
 
 @app.route('/delete',methods=['POST'])
 def delete():
     detect = sqlite3.connect("detect_results.db")
     cursor = detect.cursor()
-    paths = request.json['paths']
-    paths = [path.split(HOST)[1] for path in paths]
-    cursor.execute("""select * from tagclassified""")
-    result_list = cursor.fetchall()
-    result_dict = {tag:pickle.loads(img0s_dump) for tag,img0s_dump in result_list}
-    print(len(result_dict[0]))
+    relpaths = request.json['paths']
+    paths = [webpath_from_relpath(relpath) for relpath in relpaths]
+    print('del '+str(paths))
+    # cursor.execute("""select * from tagclassified""")
+    # result_list = cursor.fetchall()
+    # result_dict = {tag:pickle.loads(img0s_dump) for tag,img0s_dump in result_list}
     for path in paths:
         if(not os.path.exists(path)): continue
-        os.remove(path)
-        tag_names = get_tag(cursor,path)
-        tags = [names.index(tag_name) for tag_name in tag_names]
-        for tag in tags:
-            result_dict[tag].remove(path)
+        if(not path in Tag):continue
+        tag_names = get_tag(path)
 
-    print(len(result_dict[0]))
+        tags = [names.index(tag_name) for tag_name in tag_names]#转成序号
+        for tag in tags:
+            TagGroup[tag].remove(path)
+        os.remove(path)  # 文件删除
+        Tag.pop(path) #Tag表中删除
+    print('current tag group size: '+str(len(TagGroup[0])))
     if(len(paths)>0):
         s = str(paths)[1:-1]
         cursor.execute("delete from Tag where path in (" + s + ")")
-    for tag in result_dict:
-        cursor.execute("""update tagclassified set imgs = ? where tag = ?""", (pickle.dumps(result_dict[tag]),tag))
+    for tag in TagGroup:
+        cursor.execute("""update tagclassified set imgs = ? where tag = ?""", (pickle.dumps(TagGroup[tag]),tag))
     detect.commit()
     detect.close()
     return 'done',200
@@ -157,7 +151,10 @@ def get_dirs():
 def import_dir():
     dir = request.json['dir']
     print(dir)
-    web_dir = dir.rsplit('/',1)[1]
+    dir_splited = dir.rsplit('/',1)[1]
+    if(len(dir_splited)==2):
+        web_dir = dir_splited[1]
+    if(web_dir=='temp'): return 'temp reserved',202
     orig_dir = web_dir
     times = 1
     while(web_dir in PathDict):#后面加数字与已出现的同名文件夹区分
@@ -169,7 +166,10 @@ def import_dir():
     print(rel_path)
     PathDict[web_dir] = rel_path
     # pre_dir(web_dir, rel_path)
-    return 'done',200
+    return {'webdir':web_dir}
+@app.route('/del_dir',methods=['POST'])
+def del_dir():
+    dir = dir = request.json['dir']
 
 
 if __name__ == '__main__':
